+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Spell slots of prepared spellbooks are now updated in real time (#137)
 - Added an option to mark a spell as a domain or school bonus spell, which will not cost any spell slots
 - Added a numeric field for spell material component costs
+- Conditional Modifiers can now be copied via Drag and Drop
 
 ### Localization